Create It's All Greek to Me!.cs
Assignment Description

Imagine you've built a time machine so you can travel back in time to meet the Greeks (specifically, Pythagoras and Hipparchus) who have helped give us amazing theorems (Pythagorean Theorem) and branches of mathematics (Trigonometry) that directly help us in our game development! Luckily, you're bringing your computer along with you (and, of course, a power source of some kind) so you can show off your programming skills.

In this assignment, you'll calculate the distance between two points and the angle a game character would need to move in to go from the second point to the first point.

Why do we care?

There are lots of times in game development where we need to know the distance between two points. For example, an NPC might determine which target to go after based on which one is closest to the NPC. What about angle? For the same example, the NPC needs to figure out how to move toward the selected target; having the angle lets us build the appropriate velocity vector for the NPC.

Learning Goals

The learning goals for this assignment are:

● Use C# and Unity documentation to effectively utilize C# and Unity engine code.

● Use Write and WriteLine methods to display output in console applications.

● Assign values to variables and change their values through mathematical operators to solve specific problems.

● Prompt for and get user numeric input required for specific calculations.

The Pythagorean Theorem

The Pythagorean Theorem tells us how to calculate the length of the hypotenuse of a right triangle. A picture (from Wikipedia) will really help here:


We know from the theorem that

{c}^{2}={a}^{2}+{b}^{2}c 
2
 =a 
2
 +b 
2
 

so if we take the square root of {a}^{2}+{b}^{2}a 
2
 +b 
2
 , we'll get {c}c, the length of the hypotenuse.

Review criterialess 
For this peer-reviewed assignment, your peers will evaluate the functionality of your code to ensure that it works correctly. Grades will be assigned using the rubrics I've provided, but there's also an area where your reviewers can give you detailed feedback about the code you wrote to solve the problem.

Step-By-Step Assignment Instructionsless 
Starting the Assignment

Create a new Console Application called ProgrammingAssignment1.

Implementation Steps

Implement your solution by completing each of the steps below:

Print a “welcome” message to the user telling them that this application will calculate the distance between two points and the angle between those points.
Prompt for and get the x value for the first point from the user (prompt for the x value, read the value, parse the value and then store the value).
To prompt for an input, you could use something like:

Console.Write("Enter first x value: ");
To read, parse, and store the value, you could use something like:

float point1X=float.Parse(Console.ReadLine());
Prompt for and get the y value for the first point and the x and y values for the second point from the user
The picture below might help you with the following steps


Calculate the delta x and delta y between the two points; the term delta is often used in science and engineering for the difference or change between two values. The subtraction order matters (just like it did in Exercise 4) when you calculate the angle in a later step , so be sure to subtract the first point values from the second point values
Calculate the distance between the two points using the Pythagorean Theorem. The distance we're calculating is just the length of the hypotenuse (c) where we calculated a (delta x) and b (delta y) in the previous step. You might want to explore the documentation for the Math class to help with this calculation; you can find that documentation at https://msdn.microsoft.com/en-us/library/system.math(v=vs.110).aspx
Calculate the angle we'd have to move in to go from point 1 to point 2 (0 degrees is directly to the right). Be careful here; we're not trying to calculate any of the angles inside the triangle. Think of it this way: your character is at the first point, facing directly to the right. What angle would you need to rotate by to be facing the second point?
HUGE HINT: The Atan2 method in the Math class can really help here

Print an appropriate message and the distance between the two points
Print an appropriate message and the angle (in degrees) between the two points
WARNING: Be sure to convert the angle from radians to degrees before printing it

You can assume the user only enters floating point numbers.

The Math Looks Too Hard To Me

The math isn't nearly as bad as it looks at first glance. To calculate the delta x and delta y, all you need to do is subtraction. To calculate the distance between the two points, you need to square the delta x and delta y (multiply them by themselves), add the squares together, and take the square root of the result (use the appropriate method in the Math class to do that part). Finally, to calculate the angle between the two points, use the Atan2 method in the Math class; you'll need to read the documentation for that method to call it appropriately, of course. That's all the math you need to do.

Taking Small Bites

Solving this problem is a big step compared to the exercises, but that's why you shouldn't try to chew it up all in one bite. Just do each step above, one at a time, in order, and you'll see that you can do it. Although there are a few things for which you'll need to read documentation or do Google searches, if you tackle the assignment one step at a time you should be fine. Building a game is the same way, by the way. If we start thinking "How am I going to program this entire game?" it can be pretty intimidating! If we think in the much smaller steps we actually implement to eventually build a game, it becomes a much more achievable task.